Update README.md
Many product or services offered usually have their peak and trough period which the management need to understand so as to make effective sales maximization and cost reduction plan. This project used data from Uber between April 2014 and September 2014 and a graphical analysis to determine specific features of the car hiring system.